训练测试均通过;多角色音频名字写好就行

+++ finetune_speaker_v2.py
@@ -1,5 +1,4 @@
 import os
-import sys
 import json
 import argparse
 import itertools
@@ -99,10 +98,30 @@ def run(rank, n_gpus, hps):
   net_d = MultiPeriodDiscriminator(hps.model.use_spectral_norm).cuda(rank)
 
   # load existing model
-  _, _, _, _ = utils.load_checkpoint("./pretrained_models/G_0.pth", net_g, None, drop_speaker_emb=hps.drop_speaker_embed)
-  _, _, _, _ = utils.load_checkpoint("./pretrained_models/D_0.pth", net_d, None)
-  epoch_str = 1
-  global_step = 0
+  if hps.cont:
+      try:
+          _, _, _, epoch_str = utils.load_checkpoint(utils.latest_checkpoint_path(hps.model_dir, "G_latest.pth"), net_g, None)
+          _, _, _, epoch_str = utils.load_checkpoint(utils.latest_checkpoint_path(hps.model_dir, "D_latest.pth"), net_d, None)
+          global_step = (epoch_str - 1) * len(train_loader)
+      except:
+          print("Failed to find latest checkpoint, loading G_0.pth...")
+          if hps.train_with_pretrained_model:
+              print("Train with pretrained model...")
+              _, _, _, epoch_str = utils.load_checkpoint("./pretrained_models/G_0.pth", net_g, None)
+              _, _, _, epoch_str = utils.load_checkpoint("./pretrained_models/D_0.pth", net_d, None)
+          else:
+              print("Train without pretrained model...")
+          epoch_str = 1
+          global_step = 0
+  else:
+      if hps.train_with_pretrained_model:
+          print("Train with pretrained model...")
+          _, _, _, epoch_str = utils.load_checkpoint("./pretrained_models/G_0.pth", net_g, None)
+          _, _, _, epoch_str = utils.load_checkpoint("./pretrained_models/D_0.pth", net_d, None)
+      else:
+          print("Train without pretrained model...")
+      epoch_str = 1
+      global_step = 0
   # freeze all other layers except speaker embedding
   for p in net_g.parameters():
       p.requires_grad = True
@@ -241,20 +260,51 @@ def train_and_evaluate(rank, epoch, hps, nets, optims, schedulers, scaler, loade
 
       if global_step % hps.train.eval_interval == 0:
         evaluate(hps, net_g, eval_loader, writer_eval)
-        utils.save_checkpoint(net_g, None, hps.train.learning_rate, epoch, os.path.join(hps.model_dir, "G_{}.pth".format(global_step)))
+        
         utils.save_checkpoint(net_g, None, hps.train.learning_rate, epoch,
-                              os.path.join(hps.model_dir, "G_latest.pth".format(global_step)))
-        # utils.save_checkpoint(net_d, optim_d, hps.train.learning_rate, epoch, os.path.join(hps.model_dir, "D_{}.pth".format(global_step)))
-        old_g=os.path.join(hps.model_dir, "G_{}.pth".format(global_step-4000))
-        # old_d=os.path.join(hps.model_dir, "D_{}.pth".format(global_step-400))
-        if os.path.exists(old_g):
-          os.remove(old_g)
-        # if os.path.exists(old_d):
-        #   os.remove(old_d)
+                              os.path.join(hps.model_dir, "G_latest.pth"))
+        
+        utils.save_checkpoint(net_d, None, hps.train.learning_rate, epoch,
+                              os.path.join(hps.model_dir, "D_latest.pth"))
+        # save to google drive
+        if os.path.exists("/content/drive/MyDrive/"):
+            utils.save_checkpoint(net_g, None, hps.train.learning_rate, epoch,
+                                  os.path.join("/content/drive/MyDrive/", "G_latest.pth"))
+
+            utils.save_checkpoint(net_d, None, hps.train.learning_rate, epoch,
+                                  os.path.join("/content/drive/MyDrive/", "D_latest.pth"))
+        if hps.preserved > 0:
+          utils.save_checkpoint(net_g, None, hps.train.learning_rate, epoch,
+                                  os.path.join(hps.model_dir, "G_{}.pth".format(global_step)))
+          utils.save_checkpoint(net_d, None, hps.train.learning_rate, epoch,
+                                  os.path.join(hps.model_dir, "D_{}.pth".format(global_step)))
+          old_g = utils.oldest_checkpoint_path(hps.model_dir, "G_[0-9]*.pth",
+                                               preserved=hps.preserved)  # Preserve 4 (default) historical checkpoints.
+          old_d = utils.oldest_checkpoint_path(hps.model_dir, "D_[0-9]*.pth", preserved=hps.preserved)
+          if os.path.exists(old_g):
+            print(f"remove {old_g}")
+            os.remove(old_g)
+          if os.path.exists(old_d):
+            print(f"remove {old_d}")
+            os.remove(old_d)
+          if os.path.exists("/content/drive/MyDrive/"):
+              utils.save_checkpoint(net_g, None, hps.train.learning_rate, epoch,
+                                    os.path.join("/content/drive/MyDrive/", "G_{}.pth".format(global_step)))
+              utils.save_checkpoint(net_d, None, hps.train.learning_rate, epoch,
+                                    os.path.join("/content/drive/MyDrive/", "D_{}.pth".format(global_step)))
+              old_g = utils.oldest_checkpoint_path("/content/drive/MyDrive/", "G_[0-9]*.pth",
+                                                   preserved=hps.preserved)  # Preserve 4 (default) historical checkpoints.
+              old_d = utils.oldest_checkpoint_path("/content/drive/MyDrive/", "D_[0-9]*.pth", preserved=hps.preserved)
+              if os.path.exists(old_g):
+                  print(f"remove {old_g}")
+                  os.remove(old_g)
+              if os.path.exists(old_d):
+                  print(f"remove {old_d}")
+                  os.remove(old_d)
     global_step += 1
     if epoch > hps.max_epochs:
         print("Maximum epoch reached, closing training...")
-        sys.exit(0)
+        exit()
 
   if rank == 0:
     logger.info('====> Epoch: {}'.format(epoch))
